--- a/db/db_with_timestamp_test_util.h
+++ b/db/db_with_timestamp_test_util.h
@@ -17,7 +17,7 @@ namespace ROCKSDB_NAMESPACE {
 class DBBasicTestWithTimestampBase : public DBTestBase {
  public:
   explicit DBBasicTestWithTimestampBase(const std::string& dbname)
-      : DBTestBase(dbname, /*env_do_fsync=*/true) {}
+      : DBTestBase(dbname, /*env_do_fsync=*/false) {}
 
  protected:
   static std::string Key1(uint64_t k);