--- a/utilities/counted_fs.cc
+++ b/utilities/counted_fs.cc
@@ -11,249 +11,6 @@
 #include "rocksdb/utilities/options_type.h"
 
 namespace ROCKSDB_NAMESPACE {
-namespace {
-class CountedSequentialFile : public FSSequentialFileOwnerWrapper {
- private:
-  CountedFileSystem* fs_;
-
- public:
-  CountedSequentialFile(std::unique_ptr<FSSequentialFile>&& f,
-                        CountedFileSystem* fs)
-      : FSSequentialFileOwnerWrapper(std::move(f)), fs_(fs) {}
-
-  ~CountedSequentialFile() override { fs_->counters()->closes++; }
-
-  IOStatus Read(size_t n, const IOOptions& options, Slice* result,
-                char* scratch, IODebugContext* dbg) override {
-    IOStatus rv = target()->Read(n, options, result, scratch, dbg);
-    fs_->counters()->reads.RecordOp(rv, result->size());
-    return rv;
-  }
-
-  IOStatus PositionedRead(uint64_t offset, size_t n, const IOOptions& options,
-                          Slice* result, char* scratch,
-                          IODebugContext* dbg) override {
-    IOStatus rv =
-        target()->PositionedRead(offset, n, options, result, scratch, dbg);
-    fs_->counters()->reads.RecordOp(rv, result->size());
-    return rv;
-  }
-};
-
-class CountedRandomAccessFile : public FSRandomAccessFileOwnerWrapper {
- private:
-  CountedFileSystem* fs_;
-
- public:
-  CountedRandomAccessFile(std::unique_ptr<FSRandomAccessFile>&& f,
-                          CountedFileSystem* fs)
-      : FSRandomAccessFileOwnerWrapper(std::move(f)), fs_(fs) {}
-
-  ~CountedRandomAccessFile() override { fs_->counters()->closes++; }
-
-  IOStatus Read(uint64_t offset, size_t n, const IOOptions& options,
-                Slice* result, char* scratch,
-                IODebugContext* dbg) const override {
-    IOStatus rv = target()->Read(offset, n, options, result, scratch, dbg);
-    fs_->counters()->reads.RecordOp(rv, result->size());
-    return rv;
-  }
-
-  IOStatus MultiRead(FSReadRequest* reqs, size_t num_reqs,
-                     const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = target()->MultiRead(reqs, num_reqs, options, dbg);
-    for (size_t r = 0; r < num_reqs; r++) {
-      fs_->counters()->reads.RecordOp(reqs[r].status, reqs[r].result.size());
-    }
-    return rv;
-  }
-};
-
-class CountedWritableFile : public FSWritableFileOwnerWrapper {
- private:
-  CountedFileSystem* fs_;
-
- public:
-  CountedWritableFile(std::unique_ptr<FSWritableFile>&& f,
-                      CountedFileSystem* fs)
-      : FSWritableFileOwnerWrapper(std::move(f)), fs_(fs) {}
-
-  IOStatus Append(const Slice& data, const IOOptions& options,
-                  IODebugContext* dbg) override {
-    IOStatus rv = target()->Append(data, options, dbg);
-    fs_->counters()->writes.RecordOp(rv, data.size());
-    return rv;
-  }
-
-  IOStatus Append(const Slice& data, const IOOptions& options,
-                  const DataVerificationInfo& info,
-                  IODebugContext* dbg) override {
-    IOStatus rv = target()->Append(data, options, info, dbg);
-    fs_->counters()->writes.RecordOp(rv, data.size());
-    return rv;
-  }
-
-  IOStatus PositionedAppend(const Slice& data, uint64_t offset,
-                            const IOOptions& options,
-                            IODebugContext* dbg) override {
-    IOStatus rv = target()->PositionedAppend(data, offset, options, dbg);
-    fs_->counters()->writes.RecordOp(rv, data.size());
-    return rv;
-  }
-
-  IOStatus PositionedAppend(const Slice& data, uint64_t offset,
-                            const IOOptions& options,
-                            const DataVerificationInfo& info,
-                            IODebugContext* dbg) override {
-    IOStatus rv = target()->PositionedAppend(data, offset, options, info, dbg);
-    fs_->counters()->writes.RecordOp(rv, data.size());
-    return rv;
-  }
-
-  IOStatus Close(const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = target()->Close(options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->closes++;
-    }
-    return rv;
-  }
-
-  IOStatus Flush(const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = target()->Flush(options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->flushes++;
-    }
-    return rv;
-  }
-
-  IOStatus Sync(const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = target()->Sync(options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->syncs++;
-    }
-    return rv;
-  }
-
-  IOStatus Fsync(const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = target()->Fsync(options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->fsyncs++;
-    }
-    return rv;
-  }
-
-  IOStatus RangeSync(uint64_t offset, uint64_t nbytes, const IOOptions& options,
-                     IODebugContext* dbg) override {
-    IOStatus rv = target()->RangeSync(offset, nbytes, options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->syncs++;
-    }
-    return rv;
-  }
-};
-
-class CountedRandomRWFile : public FSRandomRWFileOwnerWrapper {
- private:
-  mutable CountedFileSystem* fs_;
-
- public:
-  CountedRandomRWFile(std::unique_ptr<FSRandomRWFile>&& f,
-                      CountedFileSystem* fs)
-      : FSRandomRWFileOwnerWrapper(std::move(f)), fs_(fs) {}
-  IOStatus Write(uint64_t offset, const Slice& data, const IOOptions& options,
-                 IODebugContext* dbg) override {
-    IOStatus rv = target()->Write(offset, data, options, dbg);
-    fs_->counters()->writes.RecordOp(rv, data.size());
-    return rv;
-  }
-
-  IOStatus Read(uint64_t offset, size_t n, const IOOptions& options,
-                Slice* result, char* scratch,
-                IODebugContext* dbg) const override {
-    IOStatus rv = target()->Read(offset, n, options, result, scratch, dbg);
-    fs_->counters()->reads.RecordOp(rv, result->size());
-    return rv;
-  }
-
-  IOStatus Flush(const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = target()->Flush(options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->flushes++;
-    }
-    return rv;
-  }
-
-  IOStatus Sync(const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = target()->Sync(options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->syncs++;
-    }
-    return rv;
-  }
-
-  IOStatus Fsync(const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = target()->Fsync(options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->fsyncs++;
-    }
-    return rv;
-  }
-
-  IOStatus Close(const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = target()->Close(options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->closes++;
-    }
-    return rv;
-  }
-};
-
-class CountedDirectory : public FSDirectoryWrapper {
- private:
-  mutable CountedFileSystem* fs_;
-  bool closed_ = false;
-
- public:
-  CountedDirectory(std::unique_ptr<FSDirectory>&& f, CountedFileSystem* fs)
-      : FSDirectoryWrapper(std::move(f)), fs_(fs) {}
-
-  IOStatus Fsync(const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = FSDirectoryWrapper::Fsync(options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->dsyncs++;
-    }
-    return rv;
-  }
-
-  IOStatus Close(const IOOptions& options, IODebugContext* dbg) override {
-    IOStatus rv = FSDirectoryWrapper::Close(options, dbg);
-    if (rv.ok()) {
-      fs_->counters()->closes++;
-      fs_->counters()->dir_closes++;
-      closed_ = true;
-    }
-    return rv;
-  }
-
-  IOStatus FsyncWithDirOptions(const IOOptions& options, IODebugContext* dbg,
-                               const DirFsyncOptions& dir_options) override {
-    IOStatus rv =
-        FSDirectoryWrapper::FsyncWithDirOptions(options, dbg, dir_options);
-    if (rv.ok()) {
-      fs_->counters()->dsyncs++;
-    }
-    return rv;
-  }
-
-  ~CountedDirectory() {
-    if (!closed_) {
-      // TODO: fix DB+CF code to use explicit Close, not rely on destructor
-      fs_->counters()->closes++;
-      fs_->counters()->dir_closes++;
-    }
-  }
-};
-}  // anonymous namespace
 
 std::string FileOpCounters::PrintCounters() const {
   std::stringstream ss;
@@ -285,16 +42,14 @@ std::string FileOpCounters::PrintCounters() const {
 }
 
 CountedFileSystem::CountedFileSystem(const std::shared_ptr<FileSystem>& base)
-    : FileSystemWrapper(base) {}
+    : InjectionFileSystem(base) {}
 
 IOStatus CountedFileSystem::NewSequentialFile(
     const std::string& f, const FileOptions& options,
     std::unique_ptr<FSSequentialFile>* r, IODebugContext* dbg) {
-  std::unique_ptr<FSSequentialFile> base;
-  IOStatus s = target()->NewSequentialFile(f, options, &base, dbg);
+  IOStatus s = InjectionFileSystem::NewSequentialFile(f, options, r, dbg);
   if (s.ok()) {
     counters_.opens++;
-    r->reset(new CountedSequentialFile(std::move(base), this));
   }
   return s;
 }
@@ -302,11 +57,9 @@ IOStatus CountedFileSystem::NewSequentialFile(
 IOStatus CountedFileSystem::NewRandomAccessFile(
     const std::string& f, const FileOptions& options,
     std::unique_ptr<FSRandomAccessFile>* r, IODebugContext* dbg) {
-  std::unique_ptr<FSRandomAccessFile> base;
-  IOStatus s = target()->NewRandomAccessFile(f, options, &base, dbg);
+  IOStatus s = InjectionFileSystem::NewRandomAccessFile(f, options, r, dbg);
   if (s.ok()) {
     counters_.opens++;
-    r->reset(new CountedRandomAccessFile(std::move(base), this));
   }
   return s;
 }
@@ -315,11 +68,9 @@ IOStatus CountedFileSystem::NewWritableFile(const std::string& f,
                                             const FileOptions& options,
                                             std::unique_ptr<FSWritableFile>* r,
                                             IODebugContext* dbg) {
-  std::unique_ptr<FSWritableFile> base;
-  IOStatus s = target()->NewWritableFile(f, options, &base, dbg);
+  IOStatus s = InjectionFileSystem::NewWritableFile(f, options, r, dbg);
   if (s.ok()) {
     counters_.opens++;
-    r->reset(new CountedWritableFile(std::move(base), this));
   }
   return s;
 }
@@ -327,11 +78,10 @@ IOStatus CountedFileSystem::NewWritableFile(const std::string& f,
 IOStatus CountedFileSystem::ReopenWritableFile(
     const std::string& fname, const FileOptions& options,
     std::unique_ptr<FSWritableFile>* result, IODebugContext* dbg) {
-  std::unique_ptr<FSWritableFile> base;
-  IOStatus s = target()->ReopenWritableFile(fname, options, &base, dbg);
+  IOStatus s =
+      InjectionFileSystem::ReopenWritableFile(fname, options, result, dbg);
   if (s.ok()) {
     counters_.opens++;
-    result->reset(new CountedWritableFile(std::move(base), this));
   }
   return s;
 }
@@ -340,12 +90,10 @@ IOStatus CountedFileSystem::ReuseWritableFile(
     const std::string& fname, const std::string& old_fname,
     const FileOptions& options, std::unique_ptr<FSWritableFile>* result,
     IODebugContext* dbg) {
-  std::unique_ptr<FSWritableFile> base;
-  IOStatus s =
-      target()->ReuseWritableFile(fname, old_fname, options, &base, dbg);
+  IOStatus s = InjectionFileSystem::ReuseWritableFile(fname, old_fname, options,
+                                                      result, dbg);
   if (s.ok()) {
     counters_.opens++;
-    result->reset(new CountedWritableFile(std::move(base), this));
   }
   return s;
 }
@@ -353,11 +101,9 @@ IOStatus CountedFileSystem::ReuseWritableFile(
 IOStatus CountedFileSystem::NewRandomRWFile(
     const std::string& name, const FileOptions& options,
     std::unique_ptr<FSRandomRWFile>* result, IODebugContext* dbg) {
-  std::unique_ptr<FSRandomRWFile> base;
-  IOStatus s = target()->NewRandomRWFile(name, options, &base, dbg);
+  IOStatus s = InjectionFileSystem::NewRandomRWFile(name, options, result, dbg);
   if (s.ok()) {
     counters_.opens++;
-    result->reset(new CountedRandomRWFile(std::move(base), this));
   }
   return s;
 }
@@ -366,12 +112,10 @@ IOStatus CountedFileSystem::NewDirectory(const std::string& name,
                                          const IOOptions& options,
                                          std::unique_ptr<FSDirectory>* result,
                                          IODebugContext* dbg) {
-  std::unique_ptr<FSDirectory> base;
-  IOStatus s = target()->NewDirectory(name, options, &base, dbg);
+  IOStatus s = InjectionFileSystem::NewDirectory(name, options, result, dbg);
   if (s.ok()) {
     counters_.opens++;
     counters_.dir_opens++;
-    result->reset(new CountedDirectory(std::move(base), this));
   }
   return s;
 }