This PR is based on RocksDB [9545](https://github.com/facebook/rocksdb/pull/9545).  This change introduces an InjectionFS class that can intercept the calls to the FileSystem and its associated File classes, allowing different behavior to be injected into the system.

Two examples of this Injection are part of this PR.  The CountedFileSystem class allows file operations (open, close, read, write, etc) to be counted.  This class existed before but takes advantage of the new infrastructure.

The NoSyncFileSystem is another InjectionFS.  This file system can be used to turn off the various sync and fsync operations.  This class is meant to be used primarily for testing.

As documented [elsewhere](https://github.com/facebook/rocksdb/issues/11035), the MacOS uses time-consuming sync operations, causing the unit tests to take a very long time.  Below are the results of a make check on a Mac showing the performance improvement:
main: make -j 10 check  5320.78s user 1995.73s system 75% cpu 2:40:45.65 total
PR: make -j 10 check  4166.23s user 1556.93s system 119% cpu 1:19:54.49 total
PR+[319](https://github.com/speedb-io/speedb/pull/319): make -j 10 check  3367.97s user 1225.99s system 209% cpu 36:36.18 total